+++ Editor/SpriteLib/SpriteLibraryPropertyString.cs
@@ -2,13 +2,13 @@ namespace UnityEditor.U2D.Animation
 {
     internal static class SpriteLibraryPropertyString
     {
-        public static string overrideEntries = "m_OverrideEntries";
-        public static string fromMain = "m_FromMain";
-        public static string spriteOverride = "m_SpriteOverride";
-        public static string sprite = "m_Sprite";
-        public static string overrideEntryCount = "m_EntryOverrideCount";
-        public static string name = "m_Name";
-        public static string hash = "m_Hash";
-        public static string categoryList = "m_CategoryList";
+        public const string overrideEntries = "m_OverrideEntries";
+        public const string fromMain = "m_FromMain";
+        public const string spriteOverride = "m_SpriteOverride";
+        public const string sprite = "m_Sprite";
+        public const string overrideEntryCount = "m_EntryOverrideCount";
+        public const string name = "m_Name";
+        public const string hash = "m_Hash";
+        public const string categoryList = "m_CategoryList";
     }
 }