+++ Runtime/SpriteLib/SpriteLibrarySourceAsset.cs
@@ -4,44 +4,49 @@ namespace UnityEngine.U2D.Animation
 {
     internal class SpriteLibrarySourceAsset : ScriptableObject
     {
+        public const string defaultName = "New Sprite Library Asset";
+        public const string extension = ".spriteLib";
+        
         [SerializeField]
-        List<SpriteLibCategoryOverride> m_Library = new List<SpriteLibCategoryOverride>();
+        List<SpriteLibCategoryOverride> m_Library = new();
+
         [SerializeField]
         string m_PrimaryLibraryGUID;
+
         [SerializeField]
         long m_ModificationHash;
+
         [SerializeField]
         int m_Version = 1;
 
         public IReadOnlyList<SpriteLibCategoryOverride> library => m_Library;
 
-        public string primaryLibraryID
-        {
-            get => m_PrimaryLibraryGUID;
-            set
-            {
-                m_PrimaryLibraryGUID = value;
-                UpdateModificationHash();
-            }
-        }
+        public string primaryLibraryGUID => m_PrimaryLibraryGUID;
 
         public long modificationHash => m_ModificationHash;
-        
         public int version => m_Version;
 
-        public void Copy(SpriteLibrarySourceAsset source)
+        public void InitializeWithAsset(SpriteLibrarySourceAsset source)
         {
-            m_Library.Clear();
-            m_Library.AddRange(source.m_Library);
+            m_Library = new List<SpriteLibCategoryOverride>(source.m_Library);
             m_PrimaryLibraryGUID = source.m_PrimaryLibraryGUID;
-            UpdateModificationHash();
+            m_ModificationHash = source.m_ModificationHash;
         }
 
         public void SetLibrary(IList<SpriteLibCategoryOverride> newLibrary)
         {
             if (!m_Library.Equals(newLibrary))
             {
-                m_Library = (List<SpriteLibCategoryOverride>)newLibrary;
+                m_Library = new List<SpriteLibCategoryOverride>(newLibrary);
+                UpdateModificationHash();
+            }
+        }
+
+        public void SetPrimaryLibraryGUID(string newPrimaryLibraryGUID)
+        {
+            if (m_PrimaryLibraryGUID != newPrimaryLibraryGUID)
+            {
+                m_PrimaryLibraryGUID = newPrimaryLibraryGUID;
                 UpdateModificationHash();
             }
         }
@@ -63,7 +68,7 @@ namespace UnityEngine.U2D.Animation
                 UpdateModificationHash();
             }
         }
-        
+
         public void RemoveCategory(int indexToRemove)
         {
             if (indexToRemove >= 0 && indexToRemove < m_Library.Count)
@@ -75,8 +80,7 @@ namespace UnityEngine.U2D.Animation
 
         void UpdateModificationHash()
         {
-            var hash = System.DateTime.Now.Ticks;
-            m_ModificationHash = hash;
-        }        
+            m_ModificationHash = SpriteLibraryUtility.GenerateHash();
+        }
     }
 }